+++ components/tailwind/solid/src/components/ui/switch.tsx
@@ -1,6 +1,6 @@
 import { Switch as ArkSwitch, type SwitchRootProps } from '@ark-ui/solid'
-import { Show, children, splitProps, type JSX } from 'solid-js'
-import { tv, type VariantProps } from 'tailwind-variants'
+import { type JSX, Show, children, splitProps } from 'solid-js'
+import { type VariantProps, tv } from 'tailwind-variants'
 
 export interface SwitchProps extends SwitchRootProps, SwitchVariantProps {
   children?: JSX.Element
@@ -10,8 +10,6 @@ export const Switch = (props: SwitchProps) => {
   const [variantProps, switchProps] = splitProps(props, ['size', 'class'])
   const [localProps, rootProps] = splitProps(switchProps, ['children'])
   const getChildren = children(() => localProps.children)
-
-  // @ts-expect-error https://github.com/nextui-org/tailwind-variants/issues/145
   const { root, control, label, thumb } = styles(variantProps)
 
   return (