+++ plugins/panda/src/theme/semantic-tokens/index.ts
@@ -5,11 +5,23 @@ import { createRadiiTokens } from './radii'
 import { shadows } from './shadows'
 
 export const createSemanticTokens = (options: PresetOptions) => {
-  const { accentColor = 'neutral', grayColor = 'neutral', borderRadius = 'sm' } = options
+  const {
+    accentColor = 'neutral',
+    additionalColors = [],
+    borderRadius = 'sm',
+    grayColor = 'neutral',
+  } = options
+
+  const availableColors = [grayColor, accentColor, ...additionalColors]
+  const extraColors = Object.fromEntries(
+    Object.entries(colors).filter(
+      ([color]) => availableColors.includes('*') || availableColors.includes(color),
+    ),
+  )
 
   return defineSemanticTokens({
     colors: {
-      ...colors,
+      ...extraColors,
       gray: colors[grayColor],
       accent: colors[accentColor],
       bg: {