+++ plugins/panda/src/types.ts
@@ -11,7 +11,15 @@ export interface PresetOptions {
    * @default 'sm'
    */
   borderRadius?: BorderRadius
+  /**
+   * Defines additional colors to be included in the preset. Use '*' to include all colors.
+   * By default, only the gray and accent colors are included.
+   * @default [] - This means no additional colors are included unless specified.
+   */
+  additionalColors?: Color[] | (string[] & {})
 }
+
+export type Color = AccentColor | GrayColor
 
 export type BorderRadius = (typeof borderRadii)[number]
 export const borderRadii = ['none', 'xs', 'sm', 'md', 'lg', 'xl', '2xl'] as const